+++ packages/aware-components/src/utils/a11y/audio/index.ts
@@ -1,14 +1,13 @@
+import { AUDIO } from '../../../constants';
 import { messages } from '../../messages';
+import { checkAutoPlay, checkControls, checkLoop } from '../media';
 
-type AudioProps = React.DetailedHTMLProps<
+export type AudioProps = React.DetailedHTMLProps<
   React.AudioHTMLAttributes<HTMLAudioElement>,
   HTMLAudioElement
 >;
 
 const TRANSCRIPT_KEYWORDS = ['transcript', 'text', 'description', 'summary'];
-
-const hasVisibleControls = (props: AudioProps) =>
-  props.controls && props.style?.display !== 'none';
 
 const hasAriaDescribedBy = (element: HTMLAudioElement) =>
   element.getAttribute('aria-describedby');
@@ -20,17 +19,6 @@ const hasTranscriptSibling = ({ parentNode }: HTMLAudioElement) =>
     )
   );
 
-export const checkAutoPlay = (props: AudioProps): string | null =>
-  props.autoPlay && !props.muted ? messages.audio.autoplay : null;
-
-export const checkControls = (props: AudioProps): string | null =>
-  !hasVisibleControls(props) ? messages.audio.controls : null;
-
-export const checkLoop = (props: AudioProps): string | null =>
-  props.loop && props.autoPlay && !hasVisibleControls(props)
-    ? messages.audio.loop
-    : null;
-
 const checkTranscript = (element: HTMLAudioElement): string | null =>
   hasAriaDescribedBy(element) || hasTranscriptSibling(element)
     ? null
@@ -41,8 +29,8 @@ export const audioChecks = (
   element: HTMLAudioElement
 ): string[] =>
   [
-    checkAutoPlay(props),
-    checkControls(props),
-    checkLoop(props),
+    checkAutoPlay(props, AUDIO),
+    checkControls(props, AUDIO),
+    checkLoop(props, AUDIO),
     checkTranscript(element),
   ].filter((check) => check !== null);